+++ platforms/android/app/src/main/java/com/widget/noname/cola/listener/ExtractListener.java
@@ -1,7 +1,7 @@
 package com.widget.noname.cola.listener;
 
 public interface ExtractListener {
-    void onExtractProgress(float progress);
+    void onExtractProgress(int progress);
 
     void onExtractDone();
